Remove unnecessary form validation for email
The HTML 5 form input type of email already handles
email validation. There are cases that are allowed by
HTML5 (e.g., "a@a") which our form used to detect
and notify the user. Since these are less likely to
happen, and the logic is still not fool-proof, simplify
the code and assumptions made therein to be
more generalizable and thus more robust. This is
supported in all modern browsers, and I tested
this also in Edge to verify that it works there, too.
https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/email