Check for success before clearing form. (#326)
* Ensure XHR success before clearing the form

Thanks to @tarasyarema for discovering the fix for this.
Originally pulled from PR #315. Makes sure that the
XHR request is completed and successful before clearing
and hiding the form.

* Remove extraneous console logging

Cleans up console output. Manual logging can
be added by users themselves as desired.

* Remove old honeypot function (dead code)

Removed as part of another PR, but I forgot to
remove the unused function. It just checks if there
is a value and returns the result of that if statement,
with unnecessary logging, so clean that all up.